requirements.txt
deepchem==2.6.1
matplotlib==3.6.2
mlflow==1.26.1
pyyaml==6.0
ray[tune]==1.13.0
scikit-learn==1.1.2
scikit-optimize==0.9.0
skorch==0.11.0
tune-sklearn==0.4.3
